+++ package.json
@@ -11,12 +11,18 @@
     "url": "git+https://github.com/freckle/gen-flow-js.git"
   },
   "scripts": {
-    "build": "rm -r dist && cp -r src dist"
+    "build": "tsc -d"
   },
   "dependencies": {
     "flowgen": "^1.17.0",
     "lodash": "^4.17.21",
     "piscina": "^3.2.0"
   },
-  "packageManager": "yarn@3.4.1"
+  "packageManager": "yarn@3.4.1",
+  "devDependencies": {
+    "@types/glob": "^8.1.0",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.15.5",
+    "typescript": "^5.0.2"
+  }
 }
